Appends new tables to old tables
This code appends new tables to old tables with rbind.
The purpose is to provide you with a set of workable tables for analyses. 
This set is stored in the file all_tables_appended in capstone3.
This code removes duplicates and some NULLs.
I created a charges_all table that has basic Cost and Charge data for both old and new patients.
But note that for old charges data (charges, chargesPOA, chargesAdmitPriority), tables contain columns that provide different information than new data (GrahamOut), so I have left both old and new charge tables in the RData file all_tables_appended, for your convenience.