Create TimeSplit_centered
This file creates a time split that is centered between first date a patient arrived and the last date they were discharged. For patients with only 1 encounter, the discharge date is randomly assigned +/- 1 day from their encounter date, which randomly assigns these patients' encounters to year 1 or year 2. Then, year1 charges are calclulated and year2 charges are calculated based on that cutoff.