Create CreatePatientMatrix_with_Dx
This file adds >4000 additional variables to our patient-level matrix; each of these new variables corresponds to a single ICD-9 code. Within each of these diagnosis code columns, the values represented in each row of this new matrix represent the number of times that the patient has had that diagnosis as a primary billing diagnosis in year 1. The new, wide table is saved as patient_matrix2.